% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eh_dataframe.R
\docType{data}
\name{eh_dataframe}
\alias{eh_dataframe}
\title{Dataframe containing all data required for computation}
\format{
A data.frame with 963 rows and 18 variables:
\describe{
\item{title}{file description}
\item{row.names}{ExperimentalHub ids}
}
}
\usage{
eh_dataframe
}
\description{
Dataframe containing all data required for computation
}
\keyword{datasets}
