% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sRGES_example.R
\docType{data}
\name{sRGES_example}
\alias{sRGES_example}
\title{Data of computed example sRGEs for HCC vs liver adjacent tissues on octad.small dataset}
\format{
A tibble with 12,442 rows and 6 variables:
\describe{
\item{pert_iname}{dbl Year price was recorded}
\item{mean}{mean sRGES for obtained drug  if n>1}
\item{n}{times this drug was obtained}
\item{median}{median sRGES for drug if n>1}
\item{sd}{standart deviation for obtained drug if n>1}
\item{sRGES}{sRGES score of the drug}
}
}
\usage{
sRGES_example
}
\description{
Data of computed example sRGEs for HCC vs liver adjacent tissues on octad.small dataset
}
\keyword{datasets}
